backend/requirements.txt
django
Django==4.1.7
djangorestframework==3.14.0
psycopg2==2.9.6
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.14.0
Pillow==9.5.0
gunicorn==20.1.0
django-otp<4.4.0
psycopg2-binary
exponent-server-sdk
django-filter==2.4.0
mysqlclient
mysql-connector-python 
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client
python-dotenv


